requirements-dev.txt
flake8
black
pre-commit
numpydoc
pytest
pytest-cov
pytest-mock
Pillow
-rrequirements.txt
